Add Analysis Dashboard route
