+++ app/views/Project/AnalysisModule/styles.css
@@ -15,108 +15,6 @@
         scrollbar-gutter: stable;
     }
 
-    .summary {
-        margin: 0 var(--dui-spacing-super-large);
-        overflow: unset;
-
-        .summary-content {
-            display: flex;
-            flex-wrap: wrap;
-            padding: 0;
-            overflow: unset;
-            gap: var(--dui-spacing-large);
-            --height-charts: 320px;
-
-            .info-card-container {
-                --gap: var(--dui-spacing-large);
-                --max-item-width: 260px;
-                --min-item-width: 180px;
-                --item-per-row: 2;
-                --ipr-minus-1: calc(var(--item-per-row) - 1);
-
-                display: flex;
-                flex-basis: 20%;
-                flex-grow: 1;
-                flex-wrap: wrap;
-                justify-content: space-between;
-                min-width: calc(var(--item-per-row) * var(--min-item-width) + var(--gap) * var(--ipr-minus-1));
-                max-width: calc(var(--item-per-row) * var(--max-item-width) + var(--gap) * var(--ipr-minus-1));
-                gap: var(--gap);
-
-                >* {
-                    flex-basis: calc((100% - (var(--gap) * var(--ipr-minus-1))) / var(--item-per-row));
-                    min-width: var(--min-item-width);
-                    max-width: var(--max-item-width);
-                }
-            }
-
-            .pie-chart-container {
-                display: flex;
-                flex-basis: 10%;
-                flex-direction: column;
-                flex-grow: 1;
-                min-width: 280px;
-                max-width: 320px;
-                min-height: var(--height-charts);
-
-                .pie-chart-content {
-                    overflow: unset;
-                }
-            }
-
-            .analyses-timeline-container {
-                display: flex;
-                flex-basis: 30%;
-                flex-grow: 1;
-                min-width: 420px;
-            }
-
-            .empty-chart {
-                display: flex;
-                align-items: center;
-                flex-direction: column;
-                flex-grow: 1;
-                justify-content: center;
-                color: var(--dui-color-text-watermark);
-
-                .icon {
-                    flex-shrink: 0;
-                    padding: var(--dui-spacing-small);
-                    color: var(--dui-color-text-disabled);
-                    font-size: 10em;
-                }
-
-                .text {
-                    padding: var(--dui-spacing-small);
-                    font-size: var(--dui-font-size-large);
-                }
-            }
-        }
-    }
-    .heading-description {
-        font-size: var(--dui-font-size-large);
-    }
-
-
-    .all-analyses {
-        margin: 0 var(--dui-spacing-super-large);
-        overflow: unset;
-
-        .analyses-container {
-            display: flex;
-            flex-direction: column;
-            overflow: unset;
-
-            .analysis-list {
-                display: flex;
-                flex-direction: column;
-                flex-grow: 1;
-                padding: 0;
-                min-height: 320px;
-                gap: var(--dui-spacing-medium);
-            }
-        }
-    }
 }
 
 .legend {
